requirements.txt
absl-py==0.7.1
astor==0.8.0
cachetools==4.0.0
Cerberus==1.3.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
Eve==0.9.1
Events==0.3
Flask==1.1.1
gast==0.2.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.20.1
h5py==2.9.0
idna==2.9
imageio==2.5.0
importlib-metadata==0.18
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
llvmlite==0.31.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mock==3.0.5
networkx==2.3
numba==0.47.0
numpy==1.16.3
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.0
path.py==12.0.1
Pillow==6.0.0
protobuf==3.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.21.1
scipy==1.3.0
simplejson==3.16.0
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==0.15.4
wrapt==1.12.1
zipp==0.5.2
